add packer and raw binary operations

+++ src/main/java/io/jafka/jeos/core/response/chain/account/Key.java
@@ -9,5 +9,5 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 public class Key {
     private String key;
-    private Long weight;
+    private Integer weight;
 }
